Move EvalLoopContainer to pr_utils to avoid additional imports
