+++ src/transformers/trainer_utils.py
@@ -29,7 +29,6 @@ from typing import Any, Dict, List, NamedTuple, Optional, Tuple, Union
 
 import numpy as np
 
-from .trainer_pt_utils import nested_concat, nested_numpify
 from .utils import (
     ExplicitEnum,
     is_psutil_available,
@@ -198,58 +197,6 @@ class TrainOutput(NamedTuple):
     global_step: int
     training_loss: float
     metrics: Dict[str, float]
-
-
-class EvalLoopContainer:
-    """
-    Container to store intermediate results of evaluation loop
-
-    Args:
-        do_nested_concat (`bool`, *optional*, defaults to `True`):
-            If set to `True`, each iteration will recursively concatenate a new object containing tensors to
-            the existing stored tensors, provided that the structure of the existing object and the new one
-            are identical. If set to `False`, all newly added tensors will be stored in a list.
-        padding_index (`int`, *optional*, defaults to -100):
-            Value used to pad tensors of different shapes when `do_nested_concat=True`.
-    """
-
-    def __init__(self, do_nested_concat: bool = True, padding_index: int = -100):
-        self.do_nested_concat = do_nested_concat
-        self.padding_index = padding_index
-        self.tensors = None
-        self.arrays = None
-
-    def add(self, tensors) -> None:
-        """Add tensors to the stored objects. If `do_nested_concat=True`, the tensors will be concatenated recursively."""
-        if self.tensors is None:
-            self.tensors = tensors if self.do_nested_concat else [tensors]
-        elif self.do_nested_concat:
-            self.tensors = nested_concat(self.tensors, tensors, padding_index=self.padding_index)
-        else:
-            self.tensors.append(tensors)
-
-    def to_cpu_and_numpy(self) -> None:
-        """Move tensors in stored objects to CPU and convert them to numpy arrays."""
-
-        # Check if we have something to add, if not just return
-        if self.tensors is None:
-            return
-
-        new_arrays = nested_numpify(self.tensors)
-        if self.arrays is None:
-            self.arrays = new_arrays
-        elif self.do_nested_concat:
-            self.arrays = nested_concat(self.arrays, new_arrays, padding_index=self.padding_index)
-        else:
-            self.arrays.extend(new_arrays)
-
-        # reset device tensors after adding to cpu
-        self.tensors = None
-
-    def get_arrays(self):
-        """Returns the numpified and moved to CPU stored objects."""
-        self.to_cpu_and_numpy()
-        return self.arrays
 
 
 PREFIX_CHECKPOINT_DIR = "checkpoint"
